consolidated code and addressed comments
